update tests & stdlib
This updates most of the straightforward tests, though some are still
failing.  In particular, most related to functions are failing for
reasons unknown.  It also remains to do the slow tests, and a key
challenge here is to update the standard library.

This removes all notion of loobean / boolean from the standard library.
This has requlred in a number of boolean functions being removed, since
they no longer make sense.

There are a number of outstanding issues, however.  For example, we have
no way to denote the "bool" type when defining a logical connective.
This could be resolved by introducing a first-class notion of logical
or.  Secondly, its entirely unclear how to deal with `force-bin`