Tidied up standard library
This removes all notion of loobean / boolean from the standard library.
This has requlred in a number of boolean functions being removed, since
they no longer make sense.

There are a number of outstanding issues, however.  For example, we have
no way to denote the "bool" type when defining a logical connective.
This could be resolved by introducing a first-class notion of logical
or.  Secondly, its entirely unclear how to deal with `force-bin`