fix: error over reporting (#772)
This puts in place a straightforward fix for error over-reporting.  The
issue was that resolution stopped in a given module on the first
iteration where an error was encountered.  This could potentially leave
unfinislied declarations in that module (even when they could have been
finalised).  For the root module (where the stdlib lives) this was
particularly problematic.