fix various oustanding issues
This resolves a number of issues around return type checking and the
handling of casts.  This also adds support for writing the bool type
directly at the source level.

This puts through various fixes for "for" and "reduce" expressions.
These were preexisting issues which are only surfacing because the
ordering of typing and preprocessing has changed.  Eventually, we'll
probably put the ordering back as this provides the most useful
feedback.