feat: improved constant propagation at MIR level (#778)
* feat: improved constant propagation at MIR level

This improves the mechanism for handling constant propagation at the MIR
level, in particular in the presence of casts.

* fix norm elimination and Interval.Bitwidth()

This puts in place a fix for normalisation elimination which was broken
by the removal of casts during constant propagation.  This also puts in
place a fix for Interval.Bitwidth() which was incorrectly calculating
expected bitwidth in the negative case.