feat: support sorting in coverage command (#794)
* refactor TablePrinter => termio

This relocates the "TablePrinter" into the "termio" package.

* refactor TablePrinter => FormattedTable

This refactors the old TablePrinter to be a table of formatted text, now
called a "formatted table".  This just integrates it better with the
remainder of the termio package.  There is still quite a bit to do, but
its a start at least.

* support sorting coverage columns

This adds rudimentary support for sorting columns in a coverage report