feat: support mir equations (#776)
* implement mir.Equation

This adds support for equations at the MIR level.  This is the first
step towards eliminating normalisation expressions, however there is a
way to go.

This puts in place incomplete support for handling conditions at the
corset level.  The idea is to manage this through the if-condition since
this is never exposed directly to the user.

* simplify function resolution

This simplifies function resolution by removing any notion of
overloading, which previously still existed.  The upshot of this is that
we can now (in principle) perform preprocessing prior to type checking.
That simplifies type checking quite a lot.

This also removes strict mode + @loob types

* update tests & stdlib

This updates most of the straightforward tests, though some are still
failing.  In particular, most related to functions are failing for
reasons unknown.  It also remains to do the slow tests, and a key
challenge here is to update the standard library.

This removes all notion of loobean / boolean from the standard library.
This has requlred in a number of boolean functions being removed, since
they no longer make sense.

There are a number of outstanding issues, however.  For example, we have
no way to denote the "bool" type when defining a logical connective.
This could be resolved by introducing a first-class notion of logical
or.  Secondly, its entirely unclear how to deal with `force-bin`

* fix various oustanding issues

This resolves a number of issues around return type checking and the
handling of casts.  This also adds support for writing the bool type
directly at the source level.

This puts through various fixes for "for" and "reduce" expressions.
These were preexisting issues which are only surfacing because the
ordering of typing and preprocessing has changed.  Eventually, we'll
probably put the ordering back as this provides the most useful
feedback.

* finish updating tests

This completes updating all tests with respect to the (lack of)
loobean/boolean distinction.

* improve error reporting for invokes

This improves error reporting for symbols when the incorrect arity is
given.

* final tweaks / bug fixes

This puts through a number of final tweaks and bug fixes after compiling
the full zkevm.

* improve translation to hir.If

This puts in place an improved translation from ast.If to hir.If given
that, at this stage, HIR doesn't have good support for equations.