feat: add multiplier example (#804)
* add preliminary multiplier example

This is actually not operational yet, since it requires carry flags at a
minimum.

* Tweaking automatically generated tests.

The basic problem is that the model is never being called due to the
fact that we never generate a complete instance.  To make this useful,
we have to be able sample the entire space efficiently.

* improved sampling mechanism, but there be dragons

This puts through some big improvements to the sampling mechanism, but
sadly there remaining some big problems.  Specifically, we almost never
sample anything from the model itself.  Rather, we always find traces
which either: (1) fail the function framework; (2) exist in the padding.

To resolve this issue requires sampling from traces which always have
the correct functional framework.  Perhaps one line of padding, then
valid stamp and clocks.