fix: missing constraints (#817)
* impose restrictions on use of begin

This imposes some additional restrictions on the use of begin in order
to ensure that conditions are always "functional".  This means that
(begin ...) can always safely be translated using logical and.

* update HIR to support equations

This updates HIR to support equations and MIR to support true
disjunction and conjunction.

* support direct translation of if

This puts in place support for "direct" translations of ifs within
expressions.  These are safe, but (currently) can be more expensive than
hand-optimised equivalents.

This removes the concept that evaluating an hir.Expr can produce zero or
more values.  Instead, an hir.Expr always produces exactly one value.

* feat: simplify guards

This applies optimisations at the AIR level to simplify guards, thereby
allowing a better comparison against the proposed v1.1

This also tightens the rules around parameter types.

* tweak next/prev typing

This tweaks the typing for next / prev to ensure they can be used in a
variety of positions.