feat: relational operators (#826)
This adds support for inequalities to Corset and also to HIR/MIR.  The
next step is the most challenging: implementing them at the AIR level
using a computed column.

At this stage, compiling inequalities down to actual constraints (i.e. AIR) 
is not yet implemented.  The key challenge is around determining the
appropriate bitwidth of the delta column.  One approach here is for all
equations to have a specific bitwidth (since this is necessary for field
agnostisticity).