feat: support logical connectives in Corset (#831)
* support logical connectives in Corset

This adds support for logical connectives at the Corset, and HIR levels.
These are then compiled down to the existing MIR constraints.

* unveil wcp

This unveils the wcp module almost completely fully for the very first
time.  It highlights, however, the need for a logical not operator at
the Corset/HIR levels.

* support logical negation

This adds support for a logical negation operator.

* unveil add / bin