feat: check instructions well-formed (#852)
* add balance checks

This adds checks to ensure instructions are correctly balanced.  For
subtraction, this is somewhat tricky.

* support source maps in asm.Parser

This adds support for constructing a suitable source map for the given
set of instructions.  This means that subsequent phases can report
appropriate errors.

* support definite assignment checking

This adds support for ensuring that all instructions are reachable, and
that registers are always defined before being used.

* checking conflicting writes and missing returns

This adds support for checking that writes do not conflict, and that all
control-flow is terminated with a return.

* check outputs definitely assigned

This adds support for checking that the outputs of a given function are
definitely assigned.