feat: support vector instructions (#862)
* base implementation of vector instruction

This puts in place most of the basics for vector instructions,
including their execution.  However, we do not as yet have a translation
for them.

* refactor translation of instructions

This puts through a useful refactoring for the translation of
individual instructions which simplifies things a lot, and makes it
relatively easy to change the way control-flow framing is managed, for
example.

* move constancy management outside of instructions

This moves the handling of constancies to the compiler, rather than
relying on the individual instructions.

* fix for parsing comments

There was an ordering issue related to comment headers ";;" and the
semi-column needed for vector instructions ";".

* improve vector compilation pipeline

This puts through a large range of improvements to the vector
compilation pipeline.  This includes forwarding, conditionals, etc.

* various fixes for vectorisation

At this stage, it seems unclear how to get the max example to be
implemented in one row.