fix: name clashes for perspective columns (#868)
Columns defined in perspective can no longer clash with columns of the
same name defined in other perspectives and/or the enclosing module.
This behaviour was maintained to ensure compatibility with the original
corset tool.  However, since all components are now generated by
go-corset we are no longer constrained in this way.