feat: split micro v macro instructions (#870)
* parameterise function

This gives rise to the notion of a "MacroInstruction" and a
"MicroInstruction".

* split out macro and micro instructions

* preliminary support for debugging assembly programs

* support lowering of instructions

* replace jz/jnz with jc

This replaces the old jump-if-zero and jump-if-not-zero with a more
general form of conditional jump.  In particular, this supports
relational operators, though they currently have no obvious way to
compile down microcode.

* fixes for vectorisation

This appears to make quite a big difference now.  I think there are
still possible cases where we could do better, especially in terms of
compaction.  But, for now, its winning.

* register splitting mostly operational

Register / instruction splitting based on desired register and field
widths is now operational.  It remains, however, to implement splitting
for multiply operations.

This also refactors the pipeline for checking macro/micro instances by
introducing a generic "Program[T]" construct.

* expand macro traces into micro traces

This puts in place support for expanding macro traces into micro traces.
This is necessary for checking programs at the micro-code level.

* put in place temmporary workaround for mul