refactor check reporting mechanism
This puts in place a fairly invasive refactoring of the way in which
constraint failures are reported.  Specifically, with respect to the
the way that the table of relevant cells is displayed.  At this stage,
it still displays the underlying register names.  However, the next step
is to fix this final part.