show source column names in constraint failures
The tool now reports the true source column name in a constraint failure
(where possible), rather than showing just the register name.  In some
cases, it will default back to the register name when it cannot
determine the true source name.  For example, if there is no source
mapping information available.  Or, if we have a perspective column
with no active selector.