fix: show proper perspective names for constraint failures (#874)
* fix highlighting cells in failure reports

The mechanism for highlighting cells in failure reports was broken when
using ANSI escapes.  This is now fixed.

* refactor debug and check commands

This refactors these commands to break up their source files into
multiple files, each located in an appropriate directory (e.g. debug for
files related to the debug command).

* refactor check reporting mechanism

This puts in place a fairly invasive refactoring of the way in which
constraint failures are reported.  Specifically, with respect to the
the way that the table of relevant cells is displayed.  At this stage,
it still displays the underlying register names.  However, the next step
is to fix this final part.

* show source column names in constraint failures

The tool now reports the true source column name in a constraint failure
(where possible), rather than showing just the register name.  In some
cases, it will default back to the register name when it cannot
determine the true source name.  For example, if there is no source
mapping information available.  Or, if we have a perspective column
with no active selector.