function calls compiling to micro-assembly
The next step is to support function call execution.  The proposed
solution here is to have an "execution environment" as an interface
which can be used in different circumstances.  For example, it can
simply check a trace is valid without expanding.  Or it can expand a
trace as part of checking.

Executing function calls at the ASM level is now operational, and
exploits the interpreter.  The default strategy is to expand a trace as
necessary with additional components.

Constraint generation is, in principle, working.  The key outstanding
problem is that of filling the traces.  In particular, for the bus lines
at the call side.  Using an existing computed column is awkward because
the granularity required is fine (e.g. different calls at different
points in the same micro-instruction can have different arguments).

One obvious option here is to realise the bus lines at the uASM level,
so that e.g. the CALL instruction can manage them.  This means we don't
have a call instruction at the uASM level and, instead, have an i/o sync
or something similar.