function calls compiling to micro-assembly
The next step is to support function call execution.  The proposed
solution here is to have an "execution environment" as an interface
which can be used in different circumstances.  For example, it can
simply check a trace is valid without expanding.  Or it can expand a
trace as part of checking.