feat: conditional modules (#878)
This puts in place support for "conditional modules". That is, modules which can be enabled or disabled using a constant-time expression (e.g. `(>= EVM_FORK SHANGHAI_FORK)`).  The system works using a "remapper" which essentially allows one to delete modules, and then it takes care of remapping all module and column identifiers in whatever remains.

This adds a test for conditional modules which simply checks that the given trace does not need to have columns for missing modules.