implement bus lines at function level
This changes the way that bus lines are handled.  Instead of being
handled only within the compiler, they now exist at the macro assembly
level as distinct registers.  This has two benefits: register splitting
is handle automatically; and, we get trace filling for free.

lookups arising from bus constraints are now included.  An outstanding
issue is that trace expansion does not work correctly in the presence of
multiple functions.  A poorman's workaround is to include traces for all
functions in the trace, not just the top-level.

These are now generating the correct lookups and constraints.  Trace
expansion, however, is not yet operating correctly.  Its not entirely
clear to me how best to resolve that.