feat: support function call syntax (#884)
* initial parser for function calls

This updates the parser to support function call syntax, and introduces
a (largely incomplete) micro instruction for representing function
calls.  It remains to introduce the notion of a BusID.

* function calls compiling to micro-assembly

The next step is to support function call execution.  The proposed
solution here is to have an "execution environment" as an interface
which can be used in different circumstances.  For example, it can
simply check a trace is valid without expanding.  Or it can expand a
trace as part of checking.

Executing function calls at the ASM level is now operational, and
exploits the interpreter.  The default strategy is to expand a trace as
necessary with additional components.

Constraint generation is, in principle, working.  The key outstanding
problem is that of filling the traces.  In particular, for the bus lines
at the call side.  Using an existing computed column is awkward because
the granularity required is fine (e.g. different calls at different
points in the same micro-instruction can have different arguments).

One obvious option here is to realise the bus lines at the uASM level,
so that e.g. the CALL instruction can manage them.  This means we don't
have a call instruction at the uASM level and, instead, have an i/o sync
or something similar.

* implement bus lines at function level

This changes the way that bus lines are handled.  Instead of being
handled only within the compiler, they now exist at the macro assembly
level as distinct registers.  This has two benefits: register splitting
is handle automatically; and, we get trace filling for free.

lookups arising from bus constraints are now included.  An outstanding
issue is that trace expansion does not work correctly in the presence of
multiple functions.  A poorman's workaround is to include traces for all
functions in the trace, not just the top-level.

These are now generating the correct lookups and constraints.  Trace
expansion, however, is not yet operating correctly.  Its not entirely
clear to me how best to resolve that.