add complete WCP implementation
This adds a complete and tested implementation of the WCP module.  This
includes test cases generated from three sources: block_6110045 replay
test; failingMmuModexp replay test; and those unit tests in
 WcpEdgeCaseTest.

It should be noted there remains some problems with the implementation,
since it allows rows which do not involve any of the target
instructions.