Initial refactoring
This puts in place a very initial refactoring which breaks a great many
things.  The goals of the refactoring are: (1) to have a strong notion
of module at the both the schema and trace levels; (2) to share AST
representations between different IRs; (3) to merge HIR with MIR; (4) to
support schemas constructed from classical constraints, and assembly
constraints.