removed all commands except a core
All commands have been removed except a core including debug.  The goal
now is to get what we have left compiling, running debug and then
passing tests.  This will require fixing the main functionality of the
system.  As part of this, we want to better integrate the way in which
the asm files are compiled and managed.  In particular, they should
starting using the new Trace instead of the function instances we
currently have.

Once everything is back working again, then we can selectively restore
delete commands, starting with check.  At each step, the goal is try
and streamline what we have so it works effectively.