removed all commands except a core
All commands have been removed except a core including debug.  The goal
now is to get what we have left compiling, running debug and then
passing tests.  This will require fixing the main functionality of the
system.  As part of this, we want to better integrate the way in which
the asm files are compiled and managed.  In particular, they should
starting using the new Trace instead of the function instances we
currently have.

This introduces the mixed macro schema allowing Assembly programs to
coexists with Corset modules.

This renames them to Module.Register() and Module.Registers().  The
purpose of this is to be consistent with the assembly language.  It also
helps to underscore the fact that regsiters are not columns.  For
example, we can form columns by coalscing registers (i.e. register
allocation).