fix: interleaving constraints (#923)
Since trace.Context is no longer used anywhere except as part of the
Corset compiler, this has been moved into the corset package.

This generalises assignment.Computation into something reusable, and 
reworks the structure of assignments.  The primary goal here is to allow 
assignments which read from one module, and write their results to another. 
This is necessary for interleaving constraints, but also others.  In addition, 
have tried to tidy up the existing code for the assignments, to be more 
consistent and easier to follow.

This adds an interleaving constraint which is an inter-module
constraint.  Furthermore, this plugs in the native interleave
computation for assigning interleaved columns.

There are quite a lot of linting errors which are not being reported due
to the lint settings for go-corset.  It would be nice to move towards
have more default settings for the linter.