fix: missing ir node implementations (#927)
* implement missing methods

This implements various missing methods for the IR nodes, though
ValueRange() for IfZero is still outstanding.  Its not clear to me what
the best option is here, but one approach would actually be to remove
this node altogether and just compile it out from the beginning.

* fix generate command

This command was not working correctly in the case that no lisp files
were provided.

* fix missing modules at MIR

Modules without registers were not compiled down to the MIR level.

* more fixes for generate command

* support print source map attribute

This adds support for printing out the source map attribute for a given
set of constraints.

* refactor corset.SourceColumn to use RegisterRef

This fundamentally changes the source mapping information generated by
the Corset compiler.  Specifically, the source map now correctly maps
all source-level Corset columns to valid MIR register references.  This
prevents any risk of confusion over which register a source-level Corset
column refers to.

However, this change does have some flow on implications which are not
fully resolved a this time.  Specifically, for the inspector and check
error reporting.  The issue is that e.g. a given ModuleView is actually
a view over multiple potential MIR modules.  That is, since a given
source-level Corset module now maps to one *or more* MIR modules (i.e.
to handle length multipliers).

* enable subistution of externalised constants

This was not previously enabled!

* lints