assembly trace expansion partially working
At this stage, it is now possible to expand an assembly trace.  However,
there is no constraint at this stage to check.  Furthermore, expansion
is only working at the uasm level for various reasons.

The program counter register is now an explict computed register in an
io.Function, and is represented as such in io.State.  Internally,
io.State treats the PC differently from other registers.  Specifically,
it is stored as a uint rather than a BigInt for efficiency purposes.
However, it can be loaded and stored just like any other register.