generalise trace expansion as worklist algorithm
This generalise trace expansion (both sequential and parallel) to use a
worklist-style algorithm.  This means that both sequential and parallel
implementations now behave in the same way with respect to dependencies.
Also, it means that we can easily account for columns which are expanded
during trace expansion.

There was also a minor bug in trace.ArrayModule.Resize()