feat: feature flag for byte decomposition over type proofs (#952)
* support configurable BitwidthGadget

This puts in place a bitwidth gadget which can be configured to use byte
decompositions or recurisve decompositions, depending upon whether or
not the limitless prover is available.

* test all optimisation levels

The testing framework is updated to check all optimisation levels.  This
does mean more tests will be run, unfortunately.  However, it also
ensures that any optimisation-related flags are properly tested.