preliminary support for vector access
This puts in place preliminary support for vector accesses.  These are
situations where the limbs of a variable are combined to form a whole,
specifically for the purposes of interop between lisp and assembly.

At this stage, the procedure does not yet work because we need to
actually split constraints / assignments, etc.  As such, the next step
is to target lookups (since these are initially the only form of
interop).  In particular, we need a framework for managing register
splitting.  This would include some kind of global environment which
directs the mapping from registers before to registers after.