restructure field agnosticity pipeline
The goal here is to move most of the machinery out of the asm package
and into a dedicated package which can be used by the general
constraints as well.  Furthermore, this begins a schema-wide
coordination process for determining the split of all registers in the
entire schema.  This is necessary in order to ensure the split used by
different components is always the same for the same set of registers.

At this stage, we have a small number of tests which illustrate the
pipeline as it is thus far.  However, lookups are not being properly
subdivided at this time.