feat: support mixed assembly tests (#957)
* add mixed assembly test

This adds a single mixed assembly test.

* fix check cli command

The check cli command required exactly two file arguments for reasons
unknown, and this was preventing one from passing in two or more
constraint files, etc.

* fix asm constraint checking

It is now the case that the original return value provided in a trace is
retained in the expanded trace (i.e. rather than being overwritten as
before).  This means its possible to have a constraint failure if the
computed result differs from what was specified.

* preliminary support for vector access

This puts in place preliminary support for vector accesses.  These are
situations where the limbs of a variable are combined to form a whole,
specifically for the purposes of interop between lisp and assembly.

At this stage, the procedure does not yet work because we need to
actually split constraints / assignments, etc.  As such, the next step
is to target lookups (since these are initially the only form of
interop).  In particular, we need a framework for managing register
splitting.  This would include some kind of global environment which
directs the mapping from registers before to registers after.

* restructure field agnosticity pipeline

The goal here is to move most of the machinery out of the asm package
and into a dedicated package which can be used by the general
constraints as well.  Furthermore, this begins a schema-wide
coordination process for determining the split of all registers in the
entire schema.  This is necessary in order to ensure the split used by
different components is always the same for the same set of registers.

At this stage, we have a small number of tests which illustrate the
pipeline as it is thus far.  However, lookups are not being properly
subdivided at this time.

* support vector access at MIR level

This puts in place support for vector access terms at the MIR level.
These are then lowered as one would expect into sums.  Furthermore,
register splitting now generates vector accesses rather than sums.

At this stage, lookup constraints are being flatterned in a rather naive
fashion.  Specfically, all top-level vector accesses are being inlined
as variables.

* partial support for instruction splitting

This puts back some partial support for managing instruction splitting
at the ASM level.  At this stage, there is more work to do here.