fix for unused selectors
This puts through an improvement for unused selectors.  This is done by
associating unused selectors with "unused register ids".  It works,
though its not super pretty.  It has the benefit that no constant
columns are created (especially as these cannot easily be computed
without anchors).  Hopefully, in the future, we'll have proper support
for constant columns and this will be unnecessary.