feat: support generalised register splitting (#973)
* Add tests for field agnosticity

This adds a minimal set of tests for field agnosticity.  Whilst a lot
more tests will be needed, this is enough to get started.

* replacing micro.Add/Mul/Sub with micro.Assign

The first part of generalising the register splitting algorithm is to
generalise the arithmetic micro instructions into a single assignment
primitive which accepts a polynomial on the right-hand side.

* support splitting polynomials

This puts through support for splitting the registers involved in a
polynomial to produce another (potentially more complex) polynomial.
The latter can then, in principle, be packetized.

* carry registers partial operational

Carry registers are now operational with some caveats.  The width
calcalutation is incorrect in some cases; secondly there is an infinite
loop.