fix: incorrect splitting subtraction (#986)
* initial support for signed assignments

This is not fully operational yet, but does put in place most of the
machinery required for managing signed assignments.  The challenge at
this stage is to figure out exactly how its supposed to work.

* add bit_shift example

This just adds a bit_shift example that had been lying around for a
while.  Nothing really to do with the issue at hand :)

* support signed assignments

This completes support for splitting signed assignments.  There are
currently some limitations in the algorithm which cannot (at this time)
be exploited due to restrictions enforced on the source language.
Likewise, there is not support yet for non-linear assignments.