feat: enable testing different field configurations (#988)
* support splitting trace files

This adds support for splitting trace files based on a "field
configuration".  This can be set via the command line, and or through
the testing framework.

* support --field cli flag

The `--field` cli flag indicates which field to use for compiling and
checking constraints.  There are limited options available right now,
but more will come.

* fix subtraction splitting algorithm

This puts in place a relatively simple fix for the subtraction splitting
algorithm.

* extend test traces for max and dec4

This extends the available traces for the max/dec4 benchmarks.  At this
stage, max is entering some kind of infinite loop with the GF_8209
curve.

* fix for handling expanded traces