feat: make rawcolumn field agnostic (#1003)
* Make RawColumn generic

This begins the process of converting RawColumn into RawColumn[T], such
that it can hold a field element after splitting and a big.Int before,
etc.

* convert RawColumn to RawColumn[bytes.BigEndian]

This partially converts RawColumn to RawColumn[bytes.BigEndian].  At
this stage, a lot remains to be done to make it work.

* working on max256 tests

At the moment, these tests are failing.  However, in principle, we
should be able to make them pass.

* use FieldConfig in asm.LoweringConfig

This replaces various situations where the field configuration was
defined in an ad-hoc fashion with uses of the FieldConfig struct.  This
just increases consistency.

* rename RegisterMappings and RegisterMapping

The goal here is to provide some more useful names.

* fix up tests

A number of tests were using negative values in their reject values.
Unfortunately, negative values have no meaning in a field agnostic
sense.  Therefore, instead, we've introduced field specific rejects
tests.

* implement hashBucketIterator

* fix more tests

* refactor bytes => word

* working on word array implementations

Currently, trying to figure out a suitable indexed array implementation
for big endian words.  Its a bit more complex as we don't have an upper
bound on space required.

* split Array into Array/MutArray

This splits the concept of an array into immutable and mutable parts.

* support array.Builder and word.BitArray

This adds support for the array.Builder interface, which provides a nice
separation of concerns.  Specifically, between the concern of
constructing a column versus that of accessing it.  Since columns never
change once they are constructed, this improves overall clarity and can
lead to more efficient behaviour.

* working on word.Pool implementation

This seems to be working quite well at this stage.

* heap pool is operational and threadsafe

This puts through a fix for the heap pool which is both operational and
threadsafe.  However, I am concerned about the impact of locking on
overall performance.

* fixes for HeapPool

This puts through fixes for a deadlock in HeapPool.  Overall, the trace
splitting is now accounting for a non-trivial amount of time.

* remove sanity check from StaticArray

The sanity check would be nice, but it really doesn't make sense given
how word.BigEndian is currently implemented.  Potentially in the future
it could be added back in.

* fix overflow in StaticArray.Pad()

There was a hidden overflow in this function.