feat: backport conditional lookup parser (#1016)
* support conditional lookups

This puts in place some relatively simplistic support for conditional
lookups.  Specifically, conditional lookups are parsed and then
translated directly by multiplying out the selectors (when present).
This means the generated constraints do not benefit from conditional
lookups at the prover level.

* tweak stdlib

The purpose of this is to provide primitives which are used in the
current implementation of the constraints.

* fix parser syntax error for conditional lookups

The parser was reporting a syntax error on the wrong S-expression in
some cases.

* minor fix for trace generation

Trace generation was previously instantiating the Trace class
incorrectly when a different class name was being used (e.g.
TraceLondon).

* patch trace generation

This puts through a patch for trace generation to ensure compatibility
with the latest version of go-corset.

* fix: name clashes for perspective columns (#868)

Columns defined in perspective can no longer clash with columns of the
same name defined in other perspectives and/or the enclosing module.
This behaviour was maintained to ensure compatibility with the original
corset tool.  However, since all components are now generated by
go-corset we are no longer constrained in this way.

* tweak CI workflow

* remove boolean predicates

This removes various boolean predicates, such as `remained-constant`
which would no longer appear in the constraints.

* enable PRs into v1.0-line

* Add logic test

* fixes for tests