feat: backport conditional lookups fully (#1018)
* implement LookupVector

This adds the notion of a "LookupVector" to replace the manual
representation of expressions and contexts used within a lookup.  The
purpose of this is to facilite adding selectors to help with conditional
lookups.

* support LookupVector in Corset front end.

This adds support for moving LookupVectors through the Corset front end.
For example, ensuring they are resolved, prepreprocessed, etc.

* partially complete support for conditional lookups

This adds support for true conditional lookups throughout the pipeline,
along with a command-line switch for defaulting to a legacy translation.
At this stage, however, it still remains to implement the conditional
evaluation logic.

* implement conditional lookup evaluation

This adds support for correct evaluation of conditional lookups (i.e.
which takes into account the selectors).  This uses optimised loops for
both filtered and unfiltered vectors.

* fix for LookupConstraint.Bounds

This was not correctly taking into account selector bounds.

* fix testing at different optimisation levels

This was failing because testing at a lower optimisation level was not
aligning correctly with expanded test inputs.  For now, expanded test
inputs will simple be tested at the default optimisation level.

* remove bitwidth sanity check

This was included just to see whether any selectors existed which were
not single bits.  In fact, there are some!