feat: simple splitting for lookup constraints (#1009)
* implement v.basic register splitting for lookups

This puts through a minimalistic register splitting algorithm for
lookups which will need significant improvements.  However, it does now
allow one of the agnostic tests to pass.  Furthermore, this updates the
test inputs for the first two agnostic tests.

* fixing mixed_large

This example tests the interconnect between zkasm and general lisp
constraints, where the interconnect is implemented via lookups.

* lookup padding is operational

Lookup padding arises when there is a mismatch between the bitwidth of a
given source-target pairing in a lookup.  The mismatch only matters when
it results in a differing number of limbs.  In such case, we pad with
zeros.

* support lookup splitting

This completes support for lookup splitting, and adds a number of test
cases.  Currently irregular columns cause a panic, and this is
essentially a case which needs to be resolved in the future.

* support bitwidth determination for signed terms

This adds support for determining the bitwidth of a term which can
evaluate to negative (i.e. signed) values.  This requires care in
certain situations.

* support infinite intervals

The need for infinite intervals is really only for some very unusual
edge cases.  Realistically, its just means we can fail gracefully should
those cases be encountered.